+++ include/LayoutMirror.hpp
@@ -22,7 +22,8 @@ public:
     void observeOpacity(GameObject* object, unsigned char opacity);
     void finishFor(PlayLayer* real);
     void destroyFor(PlayLayer* real);
-    void renderPlayerView(cocos2d::CCDirector* director, PlayLayer* real);
+    bool isStableGameplayScene(cocos2d::CCDirector* director, PlayLayer* real) const;
+    bool renderPlayerView(cocos2d::CCDirector* director, PlayLayer* real);
     NodeVisitAction beginNodeVisit(cocos2d::CCNode* node);
     void endNodeVisit(cocos2d::CCNode* node);
 
@@ -132,4 +133,5 @@ private:
     bool m_spatialIndexReady = false;
     bool m_renderMapReady = false;
     bool m_renderingLayout = false;
+    bool m_reportedFramebufferSkip = false;
 };